.ipynb_checkpoints/requirements-checkpoint.txt
fastapi==0.95.2
uvicorn==0.22.0
numpy==1.23.5
opencv-python==4.8.0.74
Pillow==9.5.0
matplotlib==3.7.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git
supervision==0.25.1
python-multipart==0.0.20
